requirements.txt
pandas
python-dotenv
google-api-python-client
google-auth
google-cloud-bigquery
pyarrow